+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Edge
 * Saving referer for comment, so we can get back to he commented object from a comment list
 * Admin view for managing comments
 * filter_item method for filtering lists in admin views moved here from tog_users
+* Updated template to install preferences plugins (support for users' preferences added to tog_social)
 
 0.6.0
 ----